+++ extensions/nginx/index.js
@@ -8,12 +8,21 @@ const path = require('path');
 const execa = require('execa');
 const Promise = require('bluebird');
 const template = require('lodash/template');
-const download = require('download');
-const got = require('got');
 
 const cli = require('../../lib');
 
 class NginxExtension extends cli.Extension {
+    migrations() {
+        const migrations = require('./migrations');
+
+        return [{
+            before: '1.2.0',
+            title: 'Migrating SSL certs',
+            skip: () => os.platform() !== 'linux' || !fs.existsSync(path.join(os.homedir(), '.acme.sh')),
+            task: migrations.migrateSSL.bind(this)
+        }];
+    }
+
     setup(cmd, argv) {
         // ghost setup --local, skip
         if (argv.local) {
@@ -87,6 +96,8 @@ class NginxExtension extends cli.Extension {
             return task.skip();
         }
 
+        const acme = require('./acme');
+
         const rootPath = path.resolve(ctx.instance.dir, 'system', 'nginx-root');
         const dhparamFile = '/etc/nginx/snippets/dhparam.pem';
         const sslParamsFile = '/etc/nginx/snippets/ssl-params.conf';
@@ -134,66 +145,12 @@ class NginxExtension extends cli.Extension {
             }
         }, {
             title: 'Installing acme.sh',
-            skip: (ctx) => ctx.dnsfail || fs.existsSync('/etc/letsencrypt/acme.sh'),
-            task: () => {
-                const acmeTmpDir = path.join(os.tmpdir(), 'acme.sh');
-                const acmeApiUrl = 'https://api.github.com/repos/Neilpang/acme.sh/releases/latest';
-
-                this.ui.logVerbose('ssl: creating /etc/letsencrypt directory', 'green');
-                // acme.sh creates the directory without global read permissions, so we need to make
-                // sure it has global read permissions first
-                return this.ui.sudo('mkdir -p /etc/letsencrypt').then(() => {
-                    this.ui.logVerbose('ssl: downloading acme.sh to temporary directory', 'green');
-                    return fs.emptyDir(acmeTmpDir)
-                }).then(() => got(acmeApiUrl)).then((response) => {
-                    if (response.statusCode !== 200) {
-                        // @todo: Should a specific type of error be thrown?
-                        throw new Error('Unable to query GitHub for ACME download URL');
-                    }
-
-                    try {
-                        response = JSON.parse(response.body).tarball_url;
-                    } catch (E) {
-                        throw new Error('Unable to parse GitHub response');
-                    }
-                    return download(response, acmeTmpDir, {extract: true});
-                }).then(() => {
-                    // The archive contains a single folder with the structure
-                    //  `{user}-{repo}-{commit}`, but we don't know what commit is
-                    //  from the API call. Since the dir is empty (we cleared it),
-                    //  the only thing in acmeTmpDir will be the extracted zip.
-                    const acmeCodeDir = path.resolve(acmeTmpDir, fs.readdirSync(acmeTmpDir)[0]);
-
-                    this.ui.logVerbose('ssl: installing acme.sh components', 'green');
-
-                    // Installs acme.sh into /etc/letsencrypt
-                    return this.ui.sudo('./acme.sh --install --home /etc/letsencrypt', {cwd: acmeCodeDir});
-                }).catch((error) => Promise.reject(new cli.errors.ProcessError(error)));
-            }
+            skip: (ctx) => ctx.dnsfail,
+            task: (ctx, task) => acme.install(this.ui, task)
         }, {
             title: 'Getting SSL Certificate from Let\'s Encrypt',
             skip: (ctx) => ctx.dnsfail,
-            task: () => {
-                const cmd = `/etc/letsencrypt/acme.sh --issue --home /etc/letsencrypt --domain ${parsedUrl.hostname} --webroot ${rootPath} ` +
-                `--reloadcmd "nginx -s reload" --accountemail ${argv.sslemail}${argv.sslstaging ? ' --staging' : ''}`;
-
-                return this.ui.sudo(cmd).catch((error) => {
-                    if (error.code === 2) {
-                        // error code 2 is given if a cert doesn't need to be renewed
-                        return Promise.resolve();
-                    }
-
-                    if (error.stderr.match(/Verify error:(Fetching|Invalid Response)/)) {
-                        // Domain verification failed
-                        return Promise.reject(new cli.errors.SystemError(
-                            'Your domain name is not pointing to the correct IP address of your server, please update it and run `ghost setup ssl` again'
-                        ));
-                    }
-
-                    // It's not an error we expect might happen, throw a ProcessError instead.
-                    return Promise.reject(new cli.errors.ProcessError(error));
-                });
-            }
+            task: () => acme.generate(this.ui, parsedUrl.hostname, rootPath, argv.sslemail, argv.sslstaging)
         }, {
             title: 'Generating Encryption Key (may take a few minutes)',
             skip: (ctx) => ctx.dnsfail || fs.existsSync(dhparamFile),